+++ tools/go/scell_context_string_echo.go
@@ -1,49 +1,72 @@
+/*
+ * Copyright (c) 2017 Cossack Labs Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package main
 
 import (
 	"encoding/base64"
 	"fmt"
-	"github.com/cossacklabs/themis/gothemis/cell"
 	"os"
+
+	"github.com/cossacklabs/themis/gothemis/cell"
+	"github.com/cossacklabs/themis/gothemis/keys"
 )
 
 func main() {
 	args := os.Args
 
 	if len(args) != 5 {
-		fmt.Printf("usage %s <command: enc | dec > <key> <message> <context>\n", os.Args[0])
+		fmt.Printf("usage:\n\t%s {enc|dec} <key> <message> <context>\n", os.Args[0])
 		os.Exit(1)
 	}
 
 	command := args[1]
 	key := args[2]
 	message := args[3]
-	context := []byte(args[4])
+	context := args[4]
 
-	sc := cell.New([]byte(key), cell.ModeContextImprint)
+	sc, err := cell.ContextImprintWithKey(&keys.SymmetricKey{Value: []byte(key)})
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Invalid key: \"%s\"", key)
+		os.Exit(1)
+	}
 
-	if "enc" == command {
-		encData, _, err := sc.Protect([]byte(message), context)
+	switch command {
+	case "enc":
+		encrypted, err := sc.Encrypt([]byte(message), []byte(context))
 		if nil != err {
 			fmt.Fprintln(os.Stderr, "Error encrypting message")
 			os.Exit(1)
 		}
-		fmt.Println(base64.StdEncoding.EncodeToString(encData))
+		fmt.Println(base64.StdEncoding.EncodeToString(encrypted))
 
-	} else if "dec" == command {
-		decodedMessage, err := base64.StdEncoding.DecodeString(message)
+	case "dec":
+		decoded, err := base64.StdEncoding.DecodeString(message)
 		if nil != err {
 			fmt.Fprintln(os.Stderr, "Error decoding message")
 			os.Exit(1)
 		}
-		decData, err := sc.Unprotect(decodedMessage, nil, context)
+		decrypted, err := sc.Decrypt(decoded, []byte(context))
 		if nil != err {
 			fmt.Fprintln(os.Stderr, "Error decrypting message")
 			os.Exit(1)
 		}
-		fmt.Println(string(decData[:]))
+		fmt.Println(string(decrypted))
 
-	} else {
+	default:
 		fmt.Fprintln(os.Stderr, "Wrong command, use \"enc\" or \"dec\"")
 		os.Exit(1)
 	}